Fix locale select dropdown sorting
Prior to this change, the list of locales was being sorted by converting
to lower case and then coerced to numbers. This transformations is not
intl compatible because it presumes ASCII centric logic.

This change ises the browsers INTL support for localeCompare and it
sorts by the users picked language.

The tests also now reflect this change.